Move map into migrated package as well
